chore: add outer wiring for query tx CLI
There are a couple missing pieces, but for the most part everything
lines up. Having this in place early will better inform the storage
design to back the querying interface.

In the long term we will most likely provide a Gordian-specific
transaction query, but in the immediate term we need a drop-in match.