fix: remove simapp dependency
The SDK team does not want to maintain any flexibility outside of Comet
compatibility in simapp, so we have mostly copied and pasted the
functionality into the new gcapp package.

Also update to the latest SDK commit; there were a small number of
relatively straightforward conflicts. Also temporarily disable the rest
and gRPC servers because of issues getting them to respect the
configuration.

Tests are fairly reliably failing as of this change, but they do pass
occasionally. This change was sizeable enough, so we will work on
deflaking them separately after this.